+++ src/GTP.h
@@ -55,7 +55,7 @@ extern FILE* cfg_logfile_handle;
 extern bool cfg_quiet;
 extern std::string cfg_options_str;
 extern bool cfg_benchmark;
-extern int cfg_fastmove_cnt;
+extern size_t cfg_fastmove_cnt;
 
 /*
     A list of all valid GTP2 commands is defined here:
